Fix task#409417 [Sample] Azure-Samples/key-vault-java-network-acl by lihong 202110091742
