--- src/main/java/com.microsoft.azure/Main.java
@@ -1,8 +1,0 @@
-package com.microsoft.azure;
-
-public class Main {
-    public static void main(String[] args) {
-        NetworkAclSample samples = new NetworkAclSample();
-        samples.createSampleVaultWithNetworkAcl();
-    }
-}
