+++ src/main/java/com/microsoft/azure/Main.java
@@ -1,0 +1,8 @@
+package com.microsoft.azure;
+
+public class Main {
+    public static void main(String[] args) {
+        NetworkAclSample samples = new NetworkAclSample();
+        samples.createSampleVaultWithNetworkAcl();
+    }
+}
